[TEST] Extract neutral SymbolicClient tests into test_symbolic_client.py
tests/unit/test_sanitizer.py had grown to ~670 lines but the bulk of
it didn't exercise sanitizer semantics at all — it tested pieces of
triton_viz/clients/symbolic_engine.py directly (SymbolicExpr tree
evaluation, reduce-shape rules, block-pointer construction, load/store
dtype derivation, the _range_to_iterator_constraint Z3 helper). These
tests lived there because test_sanitizer.py was the first file with the
Z3/SymbolicExpr imports already wired up, not because they belong to
the sanitizer.

Move 23 substrate tests (plus the _make_block_expr helper) into a new
tests/unit/test_symbolic_client.py. The file hosts SymbolicClient /
SymbolicExpr tests that any future symbolic client can share, with a
short module docstring saying so. No test bodies change; the cut/paste
is mechanical.

tests/unit/test_sanitizer.py keeps its sanitizer-specific content:
the factory-toggle test, _isolate_sanitizer_cfg /
_isolate_fn_symbolic_cache fixtures, _populate_for_launch helper, and
the four launch-lifecycle regressions (pre_run cache hit/miss, post_run
state preserve/clear) — those assert on sanitizer-only cache_args /
cache_grid state so they stay here.

tests/unit/test_race_detector.py is untouched; all nine of its tests
are event-capture domain logic with no neutral extraction candidates.
tests/end_to_end/ is untouched per the planning Q&A — no
test_race_detector.py exists there, so there's nothing to dedupe, and
test_sanitizer.py there is heavily OOB-specific.

Post-move line counts: test_sanitizer.py 674 -> 145; test_race_detector.py
252 unchanged; new test_symbolic_client.py 541. Total test count
invariant (70 tests before, 70 after); full suite 241 passed, 5
skipped, 0 regressions.
